tupleless impl. of trivially_copyable_tuple, passed unit tests, more generating and saving sizeof information for components usage of lighter type_list instead of std::tuple for some aliases simplified design of name-related functionality - gathered all inferred state within name_system, leaving only the significant fields inside entity_name_metas structure, also made more sense with the name-related redirecting calls inside cosmos, entity handle and name component. name_system performs all name-related logic.

+++ src/game/systems_audiovisual/wandering_pixels_system.cpp
@@ -52,7 +52,7 @@ void wandering_pixels_system::advance_wandering_pixels_for(
 	}
 
 	if (!(cache.recorded_component.reach == wandering.reach)) {
-		cache.rng = fast_randomization(cosmos.get_rng_seed_for(it));
+		cache.rng = fast_randomization(static_cast<std::size_t>(cosmos.get_rng_seed_for(it)));
 
 		for (auto& p : cache.particles) {
 			p.pos.set(
